renderBuffer implementation
Moved most of the rendering functions over to using the renderBuffer. As a result, they will now (mostly) support unicode characters.
Also made changes to the moveInput function to improve readability; it now returns 1 for a successful move and 0 for an unsuccessful move (to prevent ecsTurnSystem from advancing when trying to walk into a wall, for example).